dywa-backup/requirements.txt
psycopg2-binary==2.8.4
Faker==2.0.3
